+++ frontend/src/components/TodoList.tsx
@@ -1,22 +1,11 @@
 import TodoItem from "./Todo";
-import { useState, useEffect } from "react";
 import type { Todo } from "../types/todo";
-import { apiClient } from "../utils/apiClient";
 
-const TodoList: React.FC = () => {
-  const [todos, setTodos] = useState<Todo[]>([]);
+interface Props {
+  todos: Todo[];
+}
 
-  const fetchTodos = async () => {
-    setTodos(await apiClient.listTodos());
-  };
-
-  useEffect(() => {
-    // TODO: エラーハンドリング
-    // TODO: データの読み込み中であることを視覚的に示す。https://github.com/utibori-jp/yojitsu-tracker/pull/5/#discussion_r2093132408
-
-    fetchTodos();
-  }, []);
-
+const TodoList: React.FC<Props> = ({ todos }) => {
   if (todos.length === 0) {
     return <p className="text-center">タスクはありません。</p>;
   }
