+++ frontend/src/pages/home.tsx
@@ -3,8 +3,24 @@ import TodoList from "../components/TodoList";
 import TodoForm from "../components/TodoForm";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
+import { useState, useEffect } from "react";
+import type { Todo } from "../types/todo";
+import { apiClient } from "../utils/apiClient";
 
 const Home: React.FC = () => {
+  const [todos, setTodos] = useState<Todo[]>([]);
+  useEffect(() => {
+    // TODO: エラーハンドリング
+    const fetchTodos = async () => {
+      setTodos(await apiClient.listTodos());
+    };
+    fetchTodos();
+  }, []);
+
+  const handleUpdateTodos = (updatedTodo: Todo) => {
+    setTodos((prev) => [...prev, updatedTodo]);
+  };
+
   return (
     <main className="bg-gray-200 min-h-screen">
       <div className="bg-gray-800 p-4 shadow-md">
@@ -21,8 +37,8 @@ const Home: React.FC = () => {
 
         <TabPanel>
           <div className="space-y-4 mt-4">
-            <TodoForm />
-            <TodoList />
+            <TodoForm onSubmitSuccess={handleUpdateTodos} />
+            <TodoList todos={todos} />
           </div>
         </TabPanel>
         <TabPanel>
